dom/media/test/test_captureStream_audioLoopBack.html

<!DOCTYPE HTML>
<html>
<head>
  <title>captureStream() + loopback output check</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="../webaudio/test/webaudio.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<script>
"use strict";

add_task(async function testCaptureStreamAudioPlayout() {
  await SpecialPowers.pushPrefEnv({
    set: [
      ["media.captureStream.enabled", true],
      // For mediaDevices and getUserMedia
      ["media.navigator.permission.disabled", true],
      ["media.devices.unfocused.enabled", true],
    ],
  });

  if (!navigator.userAgent.includes("Linux")) {
    ok(true, "audio loopback is only supported on Linux");
    return;
  }

  if (!navigator.mediaDevices) {
    ok(true, "No mediaDevices, then gUM cannot have been called either");
    return;
  }

  info("opening loopback input device");
  const loopbackStream = await navigator.mediaDevices.getUserMedia({ audio: true });
  ok(loopbackStream.getAudioTracks().length === 1, "Got loopback audio track");

  info("setting up graph");
  const ac = new AudioContext();
  await ac.resume();
  const source = ac.createMediaStreamSource(loopbackStream);

  info("creating and playing audio element");
  const audio = document.createElement("audio");
  audio.src = "tone_1s_nosilence.mp3";
  audio.loop = true;
  document.body.appendChild(audio);
  ok(await audio.play().then(_ => true, _ => false), "audio started playing");

  info("capturing the audio element's media stream");
  const stream = audio.captureStream();
  is(stream.getAudioTracks().length, 1, "captured stream has an audio track");

  info("waiting for loopback audio to become audible");
  const SILENCE_RMS_THRESHOLD = 0.0;
  const AUDIBLE_RMS_THRESHOLD = 0.01;
  let curRms = await waitForRmsOverThreshold(ac, source, AUDIBLE_RMS_THRESHOLD);
  ok(curRms > 0.01, `Expected audible baseline RMS ${curRms}`);

  audio.volume = 0.3;
  info(`Lowering volume to ${audio.volume}`);
  let tempAudibleThreshold = AUDIBLE_RMS_THRESHOLD * audio.volume;
  curRms = await waitForRmsOverThreshold(ac, source, tempAudibleThreshold);
  ok(curRms > tempAudibleThreshold,
     `Expected non-silent RMS ${curRms} after volume=${audio.volume}`);

  audio.volume = 0.0;
  info(`Lowering volume to ${audio.volume}`);
  curRms = await waitForRmsEqualToThreshold(ac, source, SILENCE_RMS_THRESHOLD);
  is(curRms, SILENCE_RMS_THRESHOLD,
     `Expected silent RMS ${curRms} after volume=${audio.volume}`);

  audio.volume = 1.0;
  info(`Raising volume to ${audio.volume}`);
  curRms = await waitForRmsOverThreshold(ac, source, AUDIBLE_RMS_THRESHOLD);
  ok(curRms > AUDIBLE_RMS_THRESHOLD,
     `Expected non-silent RMS ${curRms} after volume=${audio.volume}`);

  audio.muted = true;
  info("Muting audio");
  curRms = await waitForRmsEqualToThreshold(ac, source, SILENCE_RMS_THRESHOLD);
  is(curRms, SILENCE_RMS_THRESHOLD, `Expected silent RMS ${curRms} after muting`);

  audio.muted = false;
  info("Unmuting audio");
  curRms = await waitForRmsOverThreshold(ac, source, AUDIBLE_RMS_THRESHOLD);
  ok(curRms > AUDIBLE_RMS_THRESHOLD,
     `Expected non-silent RMS ${curRms} after unmuting`);

  info("cleaning up");
  audio.pause();
  audio.remove();
  loopbackStream.getTracks().forEach(t => t.stop());
  await ac.close();
});

/**
 * Helper functions
 */

async function waitForRmsEqualToThreshold(ac, src, threshold) {
  return await waitForRmsPredicate(ac, src, threshold, rms => rms == threshold);
}

async function waitForRmsOverThreshold(ac, src, threshold) {
  return await waitForRmsPredicate(ac, src, threshold, rms => rms > threshold);
}

async function waitForRmsPredicate(ac, src, threshold, predicate, timeoutMs = 10000) {
  const start = performance.now();
  while (true) {
    const rms = await measureRMSOnceWithFreshScriptProcessor(ac, src);
    if (predicate(rms)) {
      return rms;
    }
    const elapsedTime = performance.now() - start;
    if (elapsedTime > timeoutMs) {
      throw new Error(`Timed out waitForRmsPredicate ${elapsedTime}`);
    }
  }
}

async function measureRMSOnceWithFreshScriptProcessor(ac, src) {
  // Use a new ScriptProcessorNode for each measurement, so that an event does
  // not contain samples from the past.
  const sp = ac.createScriptProcessor(2048, 1, 1);
  src.connect(sp);
  sp.connect(ac.destination);

  const { inputBuffer } = await new Promise(r => (sp.onaudioprocess = r));

  src.disconnect(sp);
  sp.disconnect();
  sp.onaudioprocess = null;
  return rms(inputBuffer); // from webaudio.js
}
</script>
</body>
</html>
